Make webpage look better
Instead of showing the data in one long list, it is now organized into
tables. Some of the backend.js code had to be changed to allow the new
display, and a styles.css file was added for css on the table
elements.

The calculator is still incomplete:
    * It is still very ugly
    * The code is very hard follow (huge function needs to be split up
    * The line endings are inconsisent
These updates will come soon.