Clean up variables and logic
There were many redundant, and unused variables in the code for the
printSuppliesNeeded function. I moved all of the variable functions to
the start of the function and tried to minimise the usage of if
statements that checked for the presence of ore2 and coal.

I also fixed some minor errors that didn't affect the user in any way,
but it was improper code. There was an extra semicolon and a break
statement after a return statmenet (the break would never be reached).

I am working on a function that shows the icon of each item next to
its name. The icons would be pulled from the RuneScape Item API. It
should be ready for the next release.