Add support for letters in budget
When you enter the budget you are now allowed to do 1k for 1,000. This
works for 'k' (thousand), 'm' (million), and 'b' (billion).

The max time setting can be ignored by the calculator if you put '0' as
your max time. It will then just make sure that it does not exceed your
budget.